Refactor task update code for invocies, use services
